Update actions.md (#366)
I think the description of standalone actions could be better. As far as I understand it, default (non-standalone) actions always act on the selected records (in index view) or the current record (in show view). Is that correct?

When I added a non-standalone action and clicked on it in the index view (without first selecting any records), nothing happened. Maybe a popup saying "please select some records first" would be more helpful?